Crate generational_box

Generational Box

Generational Box is a runtime for Rust that allows any static type to implement Copy. It can be combined with a global runtime to create an ergonomic state solution like dioxus-signals. This crate contains no unsafe code.

Three main types manage state in Generational Box:

Store: Handles recycling generational boxes that have been dropped. Your application should have one store or one store per thread.

Owner: Handles dropping generational boxes. The owner acts like a runtime lifetime guard. Any states that you create with an owner will be dropped when that owner is dropped.

GenerationalBox: The core Copy state type. The generational box will be dropped when the owner is dropped.

Example:

use generational_box::{UnsyncStorage, AnyStorage};

// Create an owner for some state for a scope
let owner = UnsyncStorage::owner();

// Create some non-copy data, move it into a owner, and work with copy data
let data: String = "hello world".to_string();
let key = owner.insert(data);

// The generational box can be read from and written to like a RefCell
let value = key.read();
assert_eq!(*value, "hello world");

How it works

Internally, generational-box creates an arena of generational RefCell’s that are recycled when the owner is dropped. You can think of the cells as something like &'static RefCell<Box<dyn Any>> with a generational check to make recycling a cell easier to debug. Then GenerationalBox’s are Copy because the &'static pointer is Copy

Structs

AlreadyBorrowedError

An error that can occur when trying to borrow a value mutably that has already been borrowed immutably.

AlreadyBorrowedMutError

An error that can occur when trying to borrow a value that has already been borrowed mutably.

GenerationalBox

The core Copy state type. The generational box will be dropped when the Owner is dropped.

GenerationalBoxId

The type erased id of a generational box.

GenerationalRef

A reference to a value in a generational box.

GenerationalRefBorrowInfo

Information about a borrow.

GenerationalRefMut

A mutable reference to a value in a generational box.

GenerationalRefMutBorrowInfo

Information about a mutable borrow.

MemoryLocation

A dynamic memory location that can be used in a generational box.

Owner

Owner: Handles dropping generational boxes. The owner acts like a runtime lifetime guard. Any states that you create with an owner will be dropped when that owner is dropped.

SyncStorage

A thread safe storage. This is slower than the unsync storage, but allows you to share the value between threads.

UnsyncStorage

A unsync storage. This is the default storage type.

ValueDroppedError

An error that can occur when trying to use a value that has been dropped.

Enums

BorrowError

An error that can occur when trying to borrow a value.

BorrowMutError

An error that can occur when trying to borrow a value mutably.

Traits

AnyStorage

A trait for any storage backing type.

Storage

A trait for a storage backing type. (RefCell, RwLock, etc.)