generational_box

Trait Storage

pub trait Storage<Data = ()>: AnyStorage + 'static {
    // Required methods
    fn try_read(
        pointer: GenerationalPointer<Self>,
    ) -> BorrowResult<Self::Ref<'static, Data>>;
    fn try_write(
        pointer: GenerationalPointer<Self>,
    ) -> BorrowMutResult<Self::Mut<'static, Data>>;
    fn new(
        value: Data,
        caller: &'static Location<'static>,
    ) -> GenerationalPointer<Self>;
    fn new_rc(
        value: Data,
        caller: &'static Location<'static>,
    ) -> GenerationalPointer<Self>;
    fn new_reference(
        inner: GenerationalPointer<Self>,
    ) -> BorrowResult<GenerationalPointer<Self>>;
    fn change_reference(
        pointer: GenerationalPointer<Self>,
        rc_pointer: GenerationalPointer<Self>,
    ) -> BorrowResult;
}

A trait for a storage backing type. (RefCell, RwLock, etc.)

Required Methods

fn try_read( pointer: GenerationalPointer<Self>, ) -> BorrowResult<Self::Ref<'static, Data>>

Try to read the value. Returns None if the value is no longer valid.

fn try_write( pointer: GenerationalPointer<Self>, ) -> BorrowMutResult<Self::Mut<'static, Data>>

Try to write the value. Returns None if the value is no longer valid.

fn new( value: Data, caller: &'static Location<'static>, ) -> GenerationalPointer<Self>

Create a new memory location. This will either create a new memory location or recycle an old one.

fn new_rc( value: Data, caller: &'static Location<'static>, ) -> GenerationalPointer<Self>

Create a new reference counted memory location. This will either create a new memory location or recycle an old one.

fn new_reference( inner: GenerationalPointer<Self>, ) -> BorrowResult<GenerationalPointer<Self>>

Reference another location if the location is valid

This method may return an error if the other box is no longer valid or it is already borrowed mutably.

fn change_reference( pointer: GenerationalPointer<Self>, rc_pointer: GenerationalPointer<Self>, ) -> BorrowResult

Change the reference a signal is pointing to

This method may return an error if the other box is no longer valid or it is already borrowed mutably.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.